Code review. Breakup the core module a bit, separating out traces, solvers, etc. Fix units, such that all solvers use the same unit conventions, guess conventions. Fix some importing conventions. Test and make sure results are as expected.
