+++ spyral/plot/__init__.py
@@ -1,0 +1,2 @@
+from .histogram import Hist1D, Hist2D, Histogrammer
+from .cut import Cut2D, CutHandler, load_cut_json, write_cut_json
