+++ spyral/run.py
@@ -3,7 +3,6 @@ from .core.workspace import Workspace
 from .phase_1 import phase_1
 from .phase_2 import phase_2
 from .phase_3 import phase_3
-from .phase_4_kalman import phase_4_kalman
 from .phase_4_interp import phase_4_interp
 from time import time
 
@@ -27,8 +26,6 @@ def run(config: Config):
             phase_3(idx, ws, config.estimate, config.detector)
 
         if config.run.do_phase4:
-            # phase_4(idx, ws, config.detector, config.solver)
-            # phase_4_kalman(idx, ws, config.detector, config.solver, nuclear_map)
             phase_4_interp(idx, ws, config.solver, config.detector, nuclear_map)
         
     stop = time()
