+++ spyral/solvers/solver_turi.py
@@ -1,34 +1,21 @@
-from .config import DetectorParameters
-from .target import Target
-from .nuclear_data import NucleusData
-from .clusterize import ClusteredCloud
-from .estimator import Direction
-from .constants import MEV_2_JOULE, MEV_2_KG
+from .guess import Guess
+from ..core.config import DetectorParameters
+from ..core.target import Target
+from ..core.nuclear_data import NucleusData
+from ..core.cluster import Cluster
+from ..core.estimator import Direction
+from ..core.constants import MEV_2_JOULE, MEV_2_KG, C, E_CHARGE
+
 from scipy import optimize, integrate, constants, linalg
 import numpy as np
 import math
-from dataclasses import dataclass
-
-@dataclass
-class InitialValue:
-    polar: float = 0.0 #radians
-    azimuthal: float = 0.0 #radians
-    brho: float = 0.0 #Tm
-    vertex_x: float = 0.0 #mm
-    vertex_y: float = 0.0 #mm
-    vertex_z: float = 0.0 #mm
-    direction: Direction = Direction.NONE 
-
-    def convert_to_array(self) -> np.ndarray:
-        return np.array([self.polar, self.azimuthal, self.brho, self.vertex_x, self.vertex_y, self.vertex_z])
+from typing import Callable
 
 TIME_WINDOW: float = 1.0e-6 #1us window
 INTERVAL: float = 1.0e-10 #0.1 ns sample
-QBRHO_2_P: float = 1.0e-9 * constants.speed_of_light #kG * cm -> MeV
 SAMPLING_PERIOD: float = 1.0e-6/512 # seconds, converts time bucket interval to time
-C: float = 3.0e8 # speed of light in m/s
 
-def motionIVP(invec: list, Efield: list, Bfield: list, dens: float, dEdx_interp: callable, q: int, m: int) -> list:
+def motionIVP(invec: list, Efield: list, Bfield: list, dens: float, dEdx_interp: Callable, q: int, m: int) -> list:
     x, y, z, vx, vy, vz = invec
     rr = np.sqrt(vx**2 + vy**2 + vz**2)
     azi = np.arctan2(vy, vx)
