+++ spyral/trace/frib_event.py
@@ -1,6 +1,6 @@
 from .frib_trace import FribTrace
 from .get_trace import Peak
-from .config import FribParameters
+from ..core.config import FribParameters
 
 import h5py as h5
 import numpy as np
